Conversational Q&A chatbot_RAG/requirements.txt
streamlit
langchain-community
langchain-core
langchain-groq
langchain-huggingface
langchain-text-splitters
langchain-chains
langchain-runnables
PyPDF2
faiss-cpu
python-dotenv
sentence-transformers